Change Dump to use dependency injection
Groupme client is passed in now instead of being created inside Dump's
constructor. In the future this will make it easier to test Dump by
mocking out said client. For now, it also makes it possible to use a
mock client for testing dump when not connected to the internet.