Refactor for dumpToChunks method
Factored the first part of the dumpFromTop method, which downloads
messages in groups of 100 and saves the groups (aka chunks) to
individual files, into a separate public method, dumpToChunks.
Having an entire group as just a folder of groups of messages
could be useful in other contexts, and besides, dumpFromTop was getting
a little long anyway.