Add exceptions.txt
This file will be used to keep up a list of errors we encounter that are
not our fault. They need to be handled somewhere at some point, so we
might as well start taking notes on what goes wrong now so we can give
better error messages later, when we actually handle them properly.