Move MAX_MESSAGES constant to GroupMessages
This is a constant enforced by GroupMe that I'm merely making note of,
moreover it pertains only to groups of messages. Moving it to be a
static property of GroupMessages makes it clearer that it's the max
number of messages we could get back in a GroupMessages object, and
not something else, like, say, the max number of messages in a group or
something.